Внесены правки по замечаниям
