+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
@@ -22,8 +23,9 @@ import static ru.yandex.practicum.filmorate.enums.OperationType.REMOVE;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-
+    @Qualifier("userDbStorage")
     private final UserStorage userStorage;
+    @Qualifier("feedDbStorage")
     private final FeedStorage feedStorage;
 
     public User register(User user) {
