+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -1,6 +1,7 @@
 package ru.yandex.practicum.filmorate.storage;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import javax.persistence.EntityNotFoundException;
@@ -9,6 +10,7 @@ import java.util.*;
 import static ru.yandex.practicum.filmorate.storage.InMemoryUserStorage.users;
 
 @Slf4j
+@Component
 public class InMemoryFilmStorage implements FilmStorage {
 
     public static final Map<Integer, Film> films = new HashMap<>();
